Fix themes object. Move component to relevant folder
